Revert "Remove ICS code from Clank Cast"
This reverts commit d889e2265a81261930ad8ecd96b929612321f57e.

Conflicts:

	chrome/android/java/src/org/chromium/chrome/browser/media/remote/LockScreenTransportControlV16.java
