+++ chrome/android/java/src/org/chromium/chrome/browser/media/remote/RemoteMediaPlayerController.java
@@ -349,13 +349,18 @@ public class RemoteMediaPlayerController implements MediaRouteController.UiListe
         return mLockScreenControl;
     }
 
+    /**
+     *
+     */
     private void createLockScreen() {
         mLockScreenControl = LockScreenTransportControl.getOrCreate(
                 mChromeVideoActivity.get(), mCurrentRouteController);
-        mLockScreenControl.setError(null);
-        mLockScreenControl.setScreenName(mCurrentRouteController.getRouteName());
-        mLockScreenControl.addListener(mCurrentRouteController);
-        mLockScreenControl.setPosterBitmap(getPoster());
+        if (mLockScreenControl != null) {
+            mLockScreenControl.setError(null);
+            mLockScreenControl.setScreenName(mCurrentRouteController.getRouteName());
+            mLockScreenControl.addListener(mCurrentRouteController);
+        }
+        if (mLockScreenControl != null) mLockScreenControl.setPosterBitmap(getPoster());
     }
 
     @Override
